Add animator ratings endpoint and test data, unify DTO formatting

+++ src/main/java/com/gladysz/kidspartymanager/dto/animator/AnimatorRatingResponseDto.java
@@ -1,9 +1,7 @@
 package com.gladysz.kidspartymanager.dto.animator;
 
-import java.math.BigDecimal;
-
 public record AnimatorRatingResponseDto(
         Long animatorId,
-        BigDecimal averageRating,
+        Double averageRating,
         long ratingsCount
-) {}
+){}
